bugfix: make email-service use same database and migrate all models
