+++ currency-rate/cmd/main.go
@@ -1,8 +1,10 @@
 package main
 
 import (
+	"context"
 	"log/slog"
 	"os"
+	"time"
 
 	rateProducer "github.com/GenesisEducationKyiv/software-engineering-school-4-0-Hukyl/currency-rate/internal/broker/rate"
 	userProducer "github.com/GenesisEducationKyiv/software-engineering-school-4-0-Hukyl/currency-rate/internal/broker/user"
@@ -24,10 +26,41 @@ const (
 	ccTo   = "UAH"
 )
 
-const (
-	rateQueueName = "rate"
-	userQueueName = "user"
+var (
+	rateQueueName = os.Getenv("RATE_QUEUE_NAME")
+	userQueueName = os.Getenv("USER_QUEUE_NAME")
 )
+
+const eventTimeout = 5 * time.Second
+
+type UserRepoDecorator struct {
+	*models.UserRepository
+	producer *userProducer.Producer
+}
+
+func (u *UserRepoDecorator) Create(user *models.User) error {
+	if err := u.UserRepository.Create(user); err != nil {
+		return err
+	}
+	ctx, cancel := context.WithTimeout(context.Background(), eventTimeout)
+	defer cancel()
+	if err := u.producer.SendSubscribe(ctx, user.Email); err != nil {
+		return err
+	}
+	return nil
+}
+
+func (u *UserRepoDecorator) Delete(user *models.User) error {
+	if err := u.UserRepository.Delete(user); err != nil {
+		return err
+	}
+	ctx, cancel := context.WithTimeout(context.Background(), eventTimeout)
+	defer cancel()
+	if err := u.producer.SendUnsubscribe(ctx, user.Email); err != nil {
+		return err
+	}
+	return nil
+}
 
 func InitDatabase() (*database.DB, error) {
 	db, err := database.New(dbCfg.NewFromEnv())
@@ -48,18 +81,24 @@ func InitFetchers() []service.RateFetcher {
 	return []service.RateFetcher{currencyBeaconFetcher, nbuFetcher}
 }
 
-func InitCron(fetcher service.RateFetcher) *cron.Manager {
-	cfg := transportCfg.NewFromEnv()
-	cfg.QueueName = rateQueueName
+func InitCron(transportConfig transportCfg.Config, fetcher service.RateFetcher) *cron.Manager {
+	transportConfig.QueueName = rateQueueName
 
-	producer, err := rateProducer.NewProducer(cfg)
+	producer, err := rateProducer.NewProducer(transportConfig)
 	if err != nil {
 		slog.Error("failed to initialize rate producer", slog.Any("error", err))
 		return nil
 	}
 
 	job := cronRate.NewCronJob(fetcher, producer, ccFrom, ccTo)
-	spec := "@every 5m"
+	spec := os.Getenv("RATE_REFRESH_CRON_SPEC")
+	if spec == "" {
+		slog.Warn(
+			"RATE_REFRESH_CRON_SPEC is not set, using default value",
+			slog.Any("default", "@every 5m"),
+		)
+		spec = "@every 5m"
+	}
 
 	cronManager := cron.NewManager()
 	cronManager.AddJob(spec, job.Run) // nolint: errcheck
@@ -81,8 +120,13 @@ func main() {
 	// Initialize rate fetcher chain of responsibilities
 	rateService := service.NewRateService(InitFetchers()...)
 
+	transportConfig := transportCfg.NewFromEnv()
+
 	// Initializer user event producer
-	userProducer, err := userProducer.NewProducer(transportCfg.NewFromEnv())
+	userProducer, err := userProducer.NewProducer(transportCfg.Config{
+		BrokerURI: transportConfig.BrokerURI,
+		QueueName: userQueueName,
+	})
 	if err != nil {
 		slog.Error("failed to initialize user producer", slog.Any("error", err))
 		panic(err)
@@ -90,7 +134,10 @@ func main() {
 	userRepo := models.NewUserRepository(db)
 
 	// Decorate userRepo with userProducer
-	decoratedUserRepo := NewUserRepoDecorator(userRepo, userProducer)
+	decoratedUserRepo := &UserRepoDecorator{
+		UserRepository: userRepo,
+		producer:       userProducer,
+	}
 
 	apiClient := server.Client{
 		Config:      serverCfg.NewFromEnv(),
@@ -98,7 +145,7 @@ func main() {
 		UserRepo:    decoratedUserRepo,
 	}
 
-	cronManager := InitCron(rateService)
+	cronManager := InitCron(transportConfig, rateService)
 	if cronManager == nil {
 		slog.Error("failed to initialize cron manager")
 	} else {
