+++ docker-compose.yml
@@ -1,5 +1,10 @@
 version: '3'
 
+x-common-variables: &compose-database-env
+  environment:
+    - DATABASE_SERVICE=postgres
+    - DATABASE_DSN=host=postgres user=postgres password=postgres dbname=genesis_kma_se_school port=5432 sslmode=disable TimeZone=UTC
+
 services:
   postgres:
     container_name: postgres
@@ -49,9 +54,7 @@ services:
     env_file:
       - ./.env
     restart: always
-    environment:
-      - DATABASE_SERVICE=postgres
-      - DATABASE_DSN=host=postgres user=postgres password=postgres dbname=genesis_kma_se_school port=5432 sslmode=disable TimeZone=UTC
+    <<: *compose-database-env
   
   email-service:
     container_name: email-service
@@ -65,10 +68,12 @@ services:
     depends_on:
       rabbitmq:
         condition: service_healthy
+      postgres:
+        condition: service_started
     env_file:
       - ./.env
     restart: always
-
+    <<: *compose-database-env
 
 volumes:
   postgres-db:
