+++ email-service/cmd/main.go
@@ -2,11 +2,15 @@ package main
 
 import (
 	"context"
+	"fmt"
 	"log/slog"
 	"os"
 	"os/signal"
 	"syscall"
+	"time"
 
+	"github.com/GenesisEducationKyiv/software-engineering-school-4-0-Hukyl/email-service/internal/broker/rate"
+	"github.com/GenesisEducationKyiv/software-engineering-school-4-0-Hukyl/email-service/internal/broker/subscriber"
 	"github.com/GenesisEducationKyiv/software-engineering-school-4-0-Hukyl/email-service/internal/mail"
 	"github.com/GenesisEducationKyiv/software-engineering-school-4-0-Hukyl/email-service/internal/mail/backends"
 	mailCfg "github.com/GenesisEducationKyiv/software-engineering-school-4-0-Hukyl/email-service/internal/mail/config"
@@ -22,12 +26,19 @@ import (
 
 const defaultCronSpec = "0 0 12 * * *"
 
+const emailTimeout = 5 * time.Second
+
+var (
+	rateQueueName = os.Getenv("RATE_QUEUE_NAME")
+	userQueueName = os.Getenv("USER_QUEUE_NAME")
+)
+
 func InitDatabase() (*database.DB, error) {
 	db, err := database.New(dbCfg.NewFromEnv())
 	if err != nil {
 		return nil, err
 	}
-	if err := db.Migrate(&models.Rate{}); err != nil {
+	if err := db.Migrate(&models.Rate{}, &models.Subscriber{}); err != nil {
 		return nil, err
 	}
 	return db, nil
@@ -35,10 +46,11 @@ func InitDatabase() (*database.DB, error) {
 
 func InitNotificationsCron(db *database.DB, mailer notifications.EmailClient) *cron.Manager {
 	// Start cron job for notifications
-	cronSpec := os.Getenv("CRON_SPEC")
+	envVar := "NOTIFICATION_CRON_SPEC"
+	cronSpec := os.Getenv(envVar)
 	if cronSpec == "" {
 		slog.Warn(
-			"CRON_SPEC is not set, using default value",
+			fmt.Sprintf("%s is not set, using default value", envVar),
 			slog.Any("default", defaultCronSpec),
 		)
 		cronSpec = defaultCronSpec
@@ -60,6 +72,100 @@ func InitNotificationsCron(db *database.DB, mailer notifications.EmailClient) *c
 	cronManager := cron.NewManager()
 	cronManager.AddJob(cronSpec, notifyF) // nolint: errcheck
 	return cronManager
+}
+
+func doWithContext(ctx context.Context, f func() error) error {
+	done := make(chan struct{})
+	go func() {
+		defer close(done)
+		if err := f(); err != nil {
+			slog.Error("error", slog.Any("error", err))
+		}
+	}()
+
+	for {
+		select {
+		case <-ctx.Done():
+			return ctx.Err()
+		case <-done:
+			return nil
+		}
+	}
+}
+
+func InitRateConsumer(config transportCfg.Config, rateRepo *models.RateRepository) *rate.Client {
+	rateConsumer, err := rate.NewClient(transportCfg.Config{
+		QueueName: rateQueueName,
+		BrokerURI: config.BrokerURI,
+	})
+	if err != nil {
+		slog.Error("creating rate client", slog.Any("error", err))
+		return nil
+	}
+	rateConsumer.Subscribe(func( // nolint: errcheck
+		ctx context.Context,
+		from, to string, rate float32, time time.Time,
+	) error {
+		slog.Info(
+			"rate fetched",
+			slog.Any("from", from),
+			slog.Any("to", to),
+			slog.Any("rate", rate),
+			slog.Any("time", time),
+		)
+		rateModel := &models.Rate{
+			CurrencyFrom: from,
+			CurrencyTo:   to,
+			Rate:         rate,
+			Created:      time.Unix(),
+		}
+		err := doWithContext(ctx, func() error {
+			return rateRepo.Create(rateModel)
+		})
+		if err != nil {
+			slog.Error("saving rate", slog.Any("error", err))
+			return err
+		}
+		return nil
+	})
+	return rateConsumer
+}
+
+func InitSubscriberConsumer(
+	config transportCfg.Config, subRepo *models.SubscriberRepository,
+) *subscriber.Client {
+	subConsumer, err := subscriber.NewClient(transportCfg.Config{
+		QueueName: userQueueName,
+		BrokerURI: config.BrokerURI,
+	})
+	if err != nil {
+		slog.Error("creating subscriber client", slog.Any("error", err))
+		return nil
+	}
+	subConsumer.SubscribeCreate(func(ctx context.Context, email string) error { // nolint: errcheck
+		slog.Info("new subscriber", slog.Any("email", email))
+		sub := &models.Subscriber{Email: email}
+		err := doWithContext(ctx, func() error {
+			return subRepo.Create(sub)
+		})
+		if err != nil {
+			slog.Error("saving subscriber", slog.Any("error", err))
+		}
+		return nil
+	})
+	subConsumer.SubscribeDelete(func(ctx context.Context, email string) error { // nolint: errcheck
+		slog.Info("delete subscriber", slog.Any("email", email))
+		subscriber := &models.Subscriber{Email: email}
+
+		err := doWithContext(ctx, func() error {
+			return subRepo.Delete(subscriber)
+		})
+		if err != nil {
+			slog.Error("deleting subscriber", slog.Any("error", err))
+		}
+		return nil
+	})
+	return subConsumer
 }
 
 func main() {
