+++ pkg/settings/.env.sample
@@ -7,11 +7,14 @@ DATABASE_SERVICE="sqlite"
 DATABASE_DSN="file::memory:?cache=shared"
 
 BROKER_URI="amqp://:@localhost:5672/"
-QUEUE_NAME="emails"
+QUEUE_NAME="default"
+RATE_QUEUE_NAME="rate"
+USER_QUEUE_NAME="user"
 BROKER_USERNAME=""
 BROKER_PASSWORD=""
 
-CRON_SPEC="0 */5 * * *"
+NOTIFICATION_CRON_SPEC="0 */10 * * *"
+RATE_REFRESH_CRON_SPEC="0 */5 * * *"
 
 CURRENCY_BEACON_API_KEY=""
